requirements.txt
fastapi
uvicorn
opencv-python
numpy
face-recognition
Pillow
psutil
pika
faiss-gpu
dlib
python-dotenv
ultralytics
pandas
ntplib
av
pickle5
pytz
logging
